show correct stats in overview chart
