+++ internal/steamclient/playerInfo.go
@@ -2,7 +2,6 @@ package steamclient
 
 import (
 	"errors"
-	"strconv"
 
 	log "github.com/sirupsen/logrus"
 )
@@ -68,37 +67,4 @@ func (sc *SteamClient) getPlayerInfo(steamID string) (PlayerInfo, error) {
 	}
 
 	return info, nil
-}
-
-func (pi PlayerInfo) LastMatch() LastMatch {
-
-	outcome := "DRAW"
-
-	if wins, err := strconv.Atoi(pi.UserStatsForGame.Stats.LastMatchWins); err == nil {
-		if rounds, err := strconv.Atoi(pi.UserStatsForGame.Stats.LastMatchRounds); err == nil {
-			if wins > (rounds / 2) {
-				outcome = "WON"
-			} else if wins < (rounds / 2) {
-				outcome = "LOST"
-			}
-		}
-	}
-
-	//TODO add correct values here!
-
-	favWeapon := getWeaponByID(pi.UserStatsForGame.Stats.LastMatchFavweaponID)
-
-	return LastMatch{
-		Outcome:           outcome,
-		FavWeaponIconPath: favWeapon.IconPath,
-		FavWeaponName:     favWeapon.Name,
-		FavWeaponAccuracy: divideStringFloats(pi.UserStatsForGame.Stats.LastMatchFavweaponHits, pi.UserStatsForGame.Stats.LastMatchFavweaponShots),
-	}
-}
-
-type LastMatch struct {
-	Outcome           string
-	FavWeaponIconPath string
-	FavWeaponName     string
-	FavWeaponAccuracy string
 }
